requirements.txt
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
pyyaml==6.0.1
pytest==8.2.0
